pyproject.toml
[tool.poetry]
name = "privategptserver"
version = "0.1.0"
description = ""
authors = ["Yossef-Dawoad <yossefdawoad15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.96.0"
langchain = "^0.0.194"
huggingface-hub = "^0.15.1"
chromadb = "^0.3.26"
poppler-utils = "^0.1.0"
sentence-transformers = "^2.2.2"
python-dotenv = "^1.0.0"
uvicorn = "^0.22.0"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
